move CommentSyntax builder inside the class
