+++ tmc-langs-java/src/main/java/fi/helsinki/cs/tmc/langs/java/AbstractJavaPlugin.java
@@ -2,6 +2,7 @@ package fi.helsinki.cs.tmc.langs.java;
 
 import fi.helsinki.cs.tmc.langs.AbstractLanguagePlugin;
 import fi.helsinki.cs.tmc.langs.abstraction.ValidationResult;
+import fi.helsinki.cs.tmc.langs.domain.CommentSyntax;
 import fi.helsinki.cs.tmc.langs.domain.CompileResult;
 import fi.helsinki.cs.tmc.langs.domain.ExerciseBuilder;
 import fi.helsinki.cs.tmc.langs.domain.ExerciseDesc;
@@ -21,7 +22,6 @@ import fi.helsinki.cs.tmc.stylerunner.exception.TMCCheckstyleException;
 import com.google.common.base.Optional;
 import com.google.common.collect.ImmutableList;
 import com.google.common.collect.ImmutableMap;
-import fi.helsinki.cs.tmc.langs.CommentSyntaxBuilder;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -52,7 +52,7 @@ public abstract class AbstractJavaPlugin extends AbstractLanguagePlugin {
     public AbstractJavaPlugin(
             Path testFolderPath, SubmissionProcessor submissionProcessor, TestScanner testScanner) {
         super(
-                new ExerciseBuilder(new CommentSyntaxBuilder()
+                new ExerciseBuilder(CommentSyntax.newBuilder()
                         .addSingleLineComment("\\/\\/")
                         .addMultiLineComment("\\/\\*+", "\\*+\\/")
                         .build()),
